+++ packages/vrender-components/src/index.ts
@@ -31,3 +31,4 @@ export * from './empty-tip';
 export * from './weather';
 export * from './util';
 export * from './label-item';
+export * from './switch';