+++ packages/vrender-core/src/plugins/builtin-plugin/edit-module.ts
@@ -1,63 +1,108 @@
 import { application } from '../../application';
 import type { IRichText, IRichTextCharacter, IRichTextParagraphCharacter } from '../../interface';
-import { IRichTextIcon, IRichTextParagraph } from '../../interface';
 
-function getMaxConfigIndexIgnoreLinebreak(textConfig: IRichTextCharacter[]) {
-  let idx = 0;
-  for (let i = 0; i < textConfig.length; i++) {
-    const c = textConfig[i] as IRichTextParagraphCharacter;
-    if (c.text !== '\n') {
-      idx++;
-    }
-  }
-  return Math.max(idx - 1, 0);
-}
+// function getMaxConfigIndexIgnoreLinebreak(textConfig: IRichTextCharacter[]) {
+//   let idx = 0;
+//   for (let i = 0; i < textConfig.length; i++) {
+//     const c = textConfig[i] as IRichTextParagraphCharacter;
+//     if (c.text !== '\n') {
+//       idx++;
+//     }
+//   }
+//   return Math.max(idx - 1, 0);
+// }
 
 /**
- * 找到cursorIndex所在的textConfig的位置，忽略换行符
+ * 找到cursorIndex所在的textConfig的位置，给出的index就是要插入的准确位置
  * @param textConfig
  * @param cursorIndex
  * @returns
  */
-export function findConfigIndex(textConfig: IRichTextCharacter[], cursorIndex: number): number {
-  let index = 0;
-  // 小于0是在最前面了
+export function findConfigIndexByCursorIdx(textConfig: IRichTextCharacter[], cursorIndex: number): number {
   if (cursorIndex < 0) {
-    return -1;
+    return 0;
   }
-  let idx = Math.round(cursorIndex);
-  for (index = 0; index < textConfig.length; index++) {
-    const c = textConfig[index] as IRichTextParagraphCharacter;
-    if (c.text !== '\n') {
-      idx--;
+
+  // 排序找到对应的元素
+  const intCursorIndex = Math.round(cursorIndex);
+  let tempCursorIndex = intCursorIndex;
+  // 跳过连续换行符中的第一个换行符
+  let lineBreak = false;
+  let configIdx = 0;
+  for (configIdx = 0; configIdx < textConfig.length && tempCursorIndex >= 0; configIdx++) {
+    const c = textConfig[configIdx] as IRichTextParagraphCharacter;
+    if (c.text === '\n') {
+      tempCursorIndex -= Number(lineBreak);
+      lineBreak = true;
+    } else {
+      tempCursorIndex--;
+      lineBreak = false;
     }
-    if (idx < 0) {
-      break;
-    }
   }
-  if (cursorIndex - Math.round(cursorIndex) < 0) {
-    index--;
+  // 说明过限了
+  if (tempCursorIndex >= 0) {
+    return textConfig.length;
   }
-  index = Math.max(index, 0);
-  // 避免超过限度，最后一个字符可能是换行符，算一个字符
-  return Math.min(index, textConfig.length - 1);
+  configIdx -= 1;
+
+  // 如果有换行，一定在换行符左边写
+  if (cursorIndex > intCursorIndex && !lineBreak) {
+    configIdx += 1;
+  }
+  return configIdx;
 }
 
-export function textConfigIgnoreLinebreakIdxToCursorIdx(textConfig: IRichTextCharacter[], cursorIndex: number): number {
-  let index = 0;
-  for (let i = 0; i < cursorIndex; i++) {
+/**
+ * 根据configIndex找到cursorIndex的位置，忽略单个换行符，连续换行符的时候只忽略第一个
+ * @param textConfig
+ * @param configIndex
+ * @returns
+ */
+export function findCursorIdxByConfigIndex(textConfig: IRichTextCharacter[], configIndex: number): number {
+  let cursorIndex = 0;
+  if (configIndex < 0) {
+    return -0.1;
+  }
+  // 仅有一个\n，那不算
+  // 如果有连续的\n，那就少算一个
+  let lastLineBreak = false;
+
+  for (let i = 0; i <= configIndex && i < textConfig.length; i++) {
     const c = textConfig[i] as IRichTextParagraphCharacter;
-    if (c.text !== '\n') {
-      index++;
+    if (c.text === '\n') {
+      cursorIndex += Number(lastLineBreak);
+      lastLineBreak = true;
+    } else {
+      cursorIndex++;
+      lastLineBreak = false;
     }
   }
-  // 正常Cursor是放在右边的，但如果回退到换行符了，那就放在左侧
-  if ((textConfig[cursorIndex] as any)?.text === '\n') {
-    index -= 0.1;
-  } else {
-    index += 0.1;
+  cursorIndex = Math.max(cursorIndex - 1, 0);
+
+  // 超出区间了直接设置到尾部，configIndex超过区间，cursorIndex不会超过
+  if (configIndex > textConfig.length - 1) {
+    // 如果最后一行是一个换行符，那么就得是xx.9否则就是xx.1
+    if ((textConfig[textConfig.length - 1] as any)?.text === '\n') {
+      return cursorIndex + 0.9;
+    }
+    return cursorIndex + 0.1;
   }
-  return index;
+
+  // 如果是这个configIdx对应到的是单个换行的话，那么算到下一个字符上
+  const lineBreak = (textConfig[configIndex] as any)?.text === '\n';
+  if (configIndex >= textConfig.length - 1 && lineBreak) {
+    return cursorIndex + 1 - 0.1;
+  }
+  const singleLineBreak = lineBreak && (textConfig[configIndex - 1] as any)?.text !== '\n';
+
+  // 光标往左放
+  cursorIndex -= 0.1;
+
+  // 如果是单行，那么这一个换行符没有算字符，光标要往右放
+  if (singleLineBreak) {
+    cursorIndex += 0.2;
+  }
+  return cursorIndex;
 }
 
 export class EditModule {
@@ -65,12 +110,16 @@ export class EditModule {
   textAreaDom: HTMLTextAreaElement;
   currRt: IRichText;
   isComposing: boolean;
+  composingConfigIdx: number;
   cursorIndex: number;
   selectionStartCursorIdx: number;
   // 输入的回调（composing的时候每次也会触发）
   onInputCbList: Array<(text: string, isComposing: boolean, cursorIdx: number, rt: IRichText) => void>;
   // change的回调（composing确认才会触发）
   onChangeCbList: Array<(text: string, isComposing: boolean, cursorIdx: number, rt: IRichText) => void>;
+  onFocusInList: Array<() => void>;
+  onFocusOutList: Array<() => void>;
+  focusOutTimer: number;
 
   constructor(container?: HTMLElement) {
     this.container = container ?? document.body;
@@ -82,8 +131,11 @@ export class EditModule {
     this.container.append(textAreaDom);
     this.textAreaDom = textAreaDom;
     this.isComposing = false;
+    this.composingConfigIdx = -1;
     this.onInputCbList = [];
     this.onChangeCbList = [];
+    this.onFocusInList = [];
+    this.onFocusOutList = [];
   }
 
   onInput(cb: (text: string, isComposing: boolean, cursorIdx: number, rt: IRichText) => void) {
@@ -92,6 +144,14 @@ export class EditModule {
 
   onChange(cb: (text: string, isComposing: boolean, cursorIdx: number, rt: IRichText) => void) {
     this.onChangeCbList.push(cb);
+  }
+
+  onFocusIn(cb: () => void) {
+    this.onFocusInList.push(cb);
+  }
+
+  onFocusOut(cb: () => void) {
+    this.onFocusOutList.push(cb);
   }
 
   applyStyle(textAreaDom: HTMLTextAreaElement) {
@@ -103,8 +163,24 @@ export class EditModule {
     textAreaDom.addEventListener('input', this.handleInput);
     textAreaDom.addEventListener('compositionstart', this.handleCompositionStart);
     textAreaDom.addEventListener('compositionend', this.handleCompositionEnd);
+    // 监听焦点
+    textAreaDom.addEventListener('focusin', this.handleFocusIn);
+    textAreaDom.addEventListener('focusout', this.handleFocusOut);
     application.global.addEventListener('keydown', this.handleKeyDown);
   }
+
+  handleFocusIn = () => {
+    // this.focusOutTimer && clearTimeout(this.focusOutTimer);
+    // this.focusOutTimer = 0;
+    // this.onFocusInList && this.onFocusInList.forEach(cb => cb());
+  };
+  handleFocusOut = () => {
+    // 暂时注释，会导致非期待情况下的误关闭
+    // // 延时触发，避免误关闭
+    // this.focusOutTimer = setTimeout(() => {
+    //   this.onFocusOutList && this.onFocusOutList.forEach(cb => cb());
+    // }, 100);
+  };
 
   handleKeyDown = (e: KeyboardEvent) => {
     if (e.key === 'Delete' || e.key === 'Backspace') {
@@ -113,49 +189,83 @@ export class EditModule {
   };
 
   handleCompositionStart = () => {
+    this.isComposing = true;
     const { textConfig = [] } = this.currRt.attribute;
+    this.composingConfigIdx = this.cursorIndex < 0 ? 0 : findConfigIndexByCursorIdx(textConfig, this.cursorIndex);
     if (this.cursorIndex < 0) {
       const config = textConfig[0];
       textConfig.unshift({ fill: 'black', ...config, text: '' });
     } else {
-      const cursorIndex = findConfigIndex(textConfig, this.cursorIndex);
-      const lastConfig = textConfig[cursorIndex];
-      textConfig.splice(cursorIndex + 1, 0, { ...lastConfig, text: '' });
+      const configIdx = this.composingConfigIdx;
+      const lastConfig = textConfig[configIdx] || textConfig[configIdx - 1];
+      textConfig.splice(configIdx, 0, { ...lastConfig, text: '' });
     }
-
-    this.isComposing = true;
   };
   handleCompositionEnd = () => {
     this.isComposing = false;
+
+    const text = this.parseCompositionStr(this.composingConfigIdx);
     // 拆分上一次的内容
+    // const { textConfig = [] } = this.currRt.attribute;
+    // const configIdx = this.composingConfigIdx;
+
+    // const lastConfig = textConfig[configIdx];
+    // textConfig.splice(configIdx, 1);
+    // const text = (lastConfig as any).text;
+    // const textList: string[] = text ? Array.from(text.toString()) : [];
+    // for (let i = 0; i < textList.length; i++) {
+    //   textConfig.splice(i + configIdx, 0, { ...lastConfig, isComposing: false, text: textList[i] } as any);
+    // }
+    // this.currRt.setAttributes({ textConfig });
+    // const nextConfigIdx = configIdx + textList.length;
+    // this.cursorIndex = findCursorIdxByConfigIndex(textConfig, nextConfigIdx);
+    this.composingConfigIdx = -1;
+
+    this.onChangeCbList.forEach(cb => {
+      cb(
+        text,
+        this.isComposing,
+        // TODO 当换行后刚开始输入会有问题，后续看这里具体Cursor变换逻辑
+        this.cursorIndex,
+        this.currRt
+      );
+    });
+  };
+
+  /**
+   * 复合输入以及粘贴，都会复制出一大段内容，这时候需要重新处理textConfig和cursorIndex
+   * 1. 拆分text到textConfig
+   * 2. 计算新的cursorIndex
+   * @param configIdx
+   */
+  parseCompositionStr(configIdx: number) {
     const { textConfig = [] } = this.currRt.attribute;
-    const configIdx = findConfigIndex(textConfig, this.cursorIndex + 1);
 
     const lastConfig = textConfig[configIdx];
     textConfig.splice(configIdx, 1);
     const text = (lastConfig as any).text;
     const textList: string[] = text ? Array.from(text.toString()) : [];
     for (let i = 0; i < textList.length; i++) {
-      textConfig.splice(i + configIdx, 0, { ...lastConfig, text: textList[i] });
+      textConfig.splice(i + configIdx, 0, {
+        fill: 'black',
+        ...lastConfig,
+        isComposing: false,
+        text: textList[i]
+      } as any);
     }
     this.currRt.setAttributes({ textConfig });
-    this.onChangeCbList.forEach(cb => {
-      cb(
-        text,
-        this.isComposing,
-        // TODO 当换行后刚开始输入会有问题，后续看这里具体Cursor变换逻辑
-        Math.min(this.cursorIndex + textList.length, getMaxConfigIndexIgnoreLinebreak(textConfig) + 0.1),
-        this.currRt
-      );
-    });
-  };
+    const nextConfigIdx = configIdx + textList.length;
+    this.cursorIndex = findCursorIdxByConfigIndex(textConfig, nextConfigIdx);
+    return text;
+  }
 
   handleInput = (ev: any) => {
     if (!this.currRt) {
       return;
     }
-
-    // 如果是回车，那就不往后+1
+    if (ev.inputType === 'historyUndo') {
+      return;
+    }
     const { textConfig = [], ...rest } = this.currRt.attribute;
     // 删完了，直接返回
     if (ev.type === 'Backspace' && !textConfig.length) {
@@ -167,77 +277,89 @@ export class EditModule {
       str = '\n';
     }
 
-    // 如果有选中多个文字，那就先删除
-    let startIdx = this.selectionStartCursorIdx;
-    let endIdx = this.cursorIndex;
-    if (startIdx > endIdx) {
-      [startIdx, endIdx] = [endIdx, startIdx];
-    }
-    // 无论是否composition都立刻恢复到没有选中的idx状态
-    this.selectionStartCursorIdx = startIdx;
-
-    // 转换成基于textConfig的
-    // let delta = 0;
-    startIdx = findConfigIndex(textConfig, startIdx);
-    // delta = this.selectionStartCursorIdx - startIdx;
-    endIdx = findConfigIndex(textConfig, endIdx);
-    // console.log(startIdx, delta, endIdx);
-
-    let idxDelta = 0;
-    // 如果是换行，得往回一格
-    if (str === '\n') {
-      idxDelta = -0.2;
+    // 处理正反选
+    if (this.selectionStartCursorIdx > this.cursorIndex) {
+      [this.cursorIndex, this.selectionStartCursorIdx] = [this.selectionStartCursorIdx, this.cursorIndex];
     }
 
-    const lastConfigIdx = startIdx + (this.isComposing ? 1 : 0);
+    const startIdx = findConfigIndexByCursorIdx(textConfig, this.selectionStartCursorIdx);
+    const endIdx = findConfigIndexByCursorIdx(textConfig, this.cursorIndex);
+
+    // composing的话会插入一个字符，所以往右加一个
+    const lastConfigIdx = this.isComposing ? this.composingConfigIdx : Math.max(startIdx - 1, 0);
+    // 算一个默认属性
     let lastConfig: any = textConfig[lastConfigIdx];
     if (!lastConfig) {
-      if (textConfig.length === 0) {
-        lastConfig = {
-          fill: rest.fill ?? 'black',
-          stroke: rest.stroke ?? false,
-          fontSize: rest.fontSize ?? 12,
-          fontWeight: rest.fontWeight ?? 'normal'
-        };
-      } else {
-        lastConfig = textConfig[lastConfigIdx - 1] || textConfig[lastConfigIdx + 1];
+      lastConfig = {
+        fill: rest.fill ?? 'black',
+        stroke: rest.stroke ?? false,
+        fontSize: rest.fontSize ?? 12,
+        fontWeight: rest.fontWeight ?? 'normal'
+      };
+    }
+    let nextConfig = lastConfig;
+
+    if (startIdx !== endIdx) {
+      textConfig.splice(startIdx, endIdx - startIdx);
+      if (this.isComposing) {
+        this.composingConfigIdx = startIdx;
       }
     }
-    let currConfig = lastConfig;
+
+    let nextConfigIdx = startIdx;
+
+    // 删除键
     if (ev.type === 'Backspace' && !this.isComposing) {
-      if (startIdx !== endIdx) {
-        textConfig.splice(startIdx + 1, endIdx - startIdx);
+      if (startIdx === endIdx) {
+        if (startIdx <= 0) {
+          return;
+        }
+        // 删除
+        textConfig.splice(startIdx - 1, 1);
+        nextConfigIdx = Math.max(startIdx - 1, 0);
       } else {
-        textConfig.splice(startIdx, 1);
-        startIdx -= 1;
+        // 不插入内容
       }
     } else {
-      if (startIdx !== endIdx) {
-        textConfig.splice(startIdx + 1, endIdx - startIdx);
-      }
-
+      // 插入
       if (!this.isComposing) {
-        currConfig = { ...lastConfig, text: '' };
-        startIdx += 1;
-        textConfig.splice(startIdx, 0, currConfig);
+        nextConfig = { fill: 'black', ...lastConfig, text: '' };
+        textConfig.splice(startIdx, 0, nextConfig);
+        nextConfigIdx++;
       }
-      (currConfig as any).text = str;
-      if (!textConfig.length) {
-        textConfig.push(currConfig);
-      }
+      // 插入
+      nextConfig.text = str;
+      // 标记isComposing，用来判定是否应该拆分成单个字符
+      nextConfig.isComposing = this.isComposing;
     }
 
     this.currRt.setAttributes({ textConfig });
-    this.cursorIndex = textConfigIgnoreLinebreakIdxToCursorIdx(textConfig, startIdx);
+    // 重新计算cursorIdx
+    // nextConfigIdx = Math.min(nextConfigIdx, textConfig.length - 1);
 
-    this.cursorIndex += idxDelta;
+    let cursorIndex = this.cursorIndex;
+    if (str && str.length > 1 && !this.isComposing) {
+      // 如果字符长度大于1且不是composing，那说明是粘贴
+      // 拆分
+      this.parseCompositionStr(nextConfigIdx - 1);
+      cursorIndex = this.cursorIndex;
+    } else {
+      // composing的时候不偏移，只有完整输入后才偏移
+      cursorIndex = findCursorIdxByConfigIndex(textConfig, nextConfigIdx);
+      if (!this.isComposing) {
+        this.cursorIndex = cursorIndex;
+      } else {
+        this.cursorIndex = this.selectionStartCursorIdx;
+      }
+    }
+
     if (!this.isComposing) {
       this.onChangeCbList.forEach(cb => {
-        cb(str, this.isComposing, this.cursorIndex, this.currRt);
+        cb(str, this.isComposing, cursorIndex, this.currRt);
       });
     } else {
       this.onInputCbList.forEach(cb => {
-        cb(str, this.isComposing, this.cursorIndex, this.currRt);
+        cb(str, this.isComposing, cursorIndex, this.currRt);
       });
     }
   };
@@ -259,6 +381,8 @@ export class EditModule {
     this.textAreaDom.removeEventListener('input', this.handleInput);
     this.textAreaDom.removeEventListener('compositionstart', this.handleCompositionStart);
     this.textAreaDom.removeEventListener('compositionend', this.handleCompositionEnd);
+    this.textAreaDom.addEventListener('focusin', this.handleFocusOut);
+    this.textAreaDom.addEventListener('focusout', this.handleFocusOut);
     application.global.removeEventListener('keydown', this.handleKeyDown);
   }
 }